Bug in HLCSTCP3:
NB: There is a recently introduced bug in HLCSTCP3 (patched up to 157 on
the second line), line 69, which says:

Q:(HLOS'["VMS")&(HLOS'["UNIX") X "U IO:(::""-M"")"

This line is incorrect in many regards: it assumes all Cache systems run
on VMS or UNIX; and it assumes that all UNIX systems will be Cache.
Neither of these assumptions are correct.